requirements.txt
scikit-learn

# use tensorflow-gpu on cuda activated machines
# uncomment librosa for when using pip, if using conda, check readme
tensorflow-gpu >= 2.1
# librosa 
numba == 0.48.0
pandas
numby
matplotlib
tqdm